requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
idna==3.4
insta-scrape==2.1.2
instaloader==4.9.5
pafy==0.5.5
Pillow==9.4.0
pytube==12.1.2
pytube3==9.6.4
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
youtube-dl==2021.12.17
